backend/requirements.txt
Django
djangorestframework
gunicorn
psycopg2-binary
django-cors-headers